Enable usage with Django 1.10's MIDDLEWARE setting
While the package was tested under 1.10, it was not tested using the new
MIDDLEWARE setting but instead the old MIDDLEWARE_CLASSES setting.

This commit:

a) forces us to use the MIDDLEWARE setting during testing under Django
1.10 and above.

b) utilises Django's MiddlewareMixin so that our Middleware can perform
across versions.

c) updates our MiddlewareMixin constructor so that it can take the
new Django supplied arguments in Django 1.10 and above.

Down the line, when MIDDLEWARE_CLASSES is official deprecated we will
have to migrate our middleware to the new format. Until then, this will
have to be our solution if we are to support old versions.